add `experimental-async` feature
