+++ newsfragments/3931.added.md
@@ -1,0 +1,1 @@
+Add `experimental-async` feature.
